Update README.md
Corrected and cleaned up grammar and sentence structure. Added details and spelled out full names of abbreviations for clarity with first usage. We should NOT assume that only people with computational backgrounds will be reading this description, in fact it is very likely the biological community of researchers will have great interest in these.